requirements.txt
# Core web framework
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.18

# PDF processing
pypdf2==3.0.1
pdf2image==1.16.3
pillow>=11.0.0

# MinerU vLLM support (optional, if using vllm-async-engine)
# Installation options:
# pip install "mineru-vl-utils[vllm]"
# Or
# pip install mineru-vl-utils vllm llguidance xgrammar

# SharePoint


# Environment variables and configuration
python-dotenv>=1.0.1

# OpenAI (compatible with MinerU 2.7.0)
openai>=1.70.0,<3

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP requests (for Ollama)
requests>=2.31.0

# Microsoft Entra ID auth helper (used for SharePoint/Graph token acquisition)
msal>=1.26.0

# Test tools (for test_websocket_progress.py; uvicorn[standard] usually includes websockets)
# Core test runner
pytest>=8.0.0
pytest-asyncio>=0.23.0
# websockets>=11.0